test(models): keep remote protocol coverage within file limits
